splitted processing from linebuilding
